fix: format spawnpoint rotations like vanilla
