+++ steel-core/src/command/builtins/setworldspawn.rs
@@ -1,6 +1,6 @@
 //! Default world spawn command.
 
-use steel_utils::{BlockPos, Identifier, translations};
+use steel_utils::{BlockPos, Identifier, java::float_to_string, translations};
 use text_components::TextComponent;
 
 use super::super::{
@@ -74,8 +74,8 @@ fn set_spawn(
             position.x().to_string(),
             position.y().to_string(),
             position.z().to_string(),
-            yaw.to_string(),
-            pitch.to_string(),
+            float_to_string(yaw),
+            float_to_string(pitch),
             source.world().key.to_string(),
         ])
         .component();
