+++ steel-utils/src/nbt/snbt/writer.rs
@@ -2,7 +2,10 @@ use std::fmt::Write as _;
 
 use simdnbt::owned::NbtTag;
 
-use crate::nbt::nbt_list_values;
+use crate::{
+    java::{double_to_string, float_to_string},
+    nbt::nbt_list_values,
+};
 
 /// Renders an NBT value with Vanilla `Tag::toString` semantics.
 ///
@@ -21,8 +24,8 @@ fn write_canonical_snbt(tag: &NbtTag, output: &mut String) -> Option<()> {
         NbtTag::Short(value) => write!(output, "{value}s").ok()?,
         NbtTag::Int(value) => write!(output, "{value}").ok()?,
         NbtTag::Long(value) => write!(output, "{value}L").ok()?,
-        NbtTag::Float(value) => write!(output, "{}f", java_float_string(*value)).ok()?,
-        NbtTag::Double(value) => write!(output, "{}d", java_double_string(*value)).ok()?,
+        NbtTag::Float(value) => write!(output, "{}f", float_to_string(*value)).ok()?,
+        NbtTag::Double(value) => write!(output, "{}d", double_to_string(*value)).ok()?,
         NbtTag::ByteArray(values) => {
             output.push_str("[B;");
             for (index, value) in values.iter().enumerate() {
@@ -133,116 +136,4 @@ fn quote_and_escape(value: &str, output: &mut String) {
         }
     }
     output.push(quote);
-}
-
-fn java_float_string(value: f32) -> String {
-    java_floating_string(
-        value.is_sign_negative(),
-        value.is_nan(),
-        value.is_infinite(),
-        value == 0.0,
-        &format!("{:.8e}", value.abs()),
-        9,
-        |precision| format!("{:.*e}", precision, value.abs()),
-        |candidate| candidate.parse::<f32>().ok().map(f32::to_bits) == Some(value.abs().to_bits()),
-    )
-}
-
-fn java_double_string(value: f64) -> String {
-    java_floating_string(
-        value.is_sign_negative(),
-        value.is_nan(),
-        value.is_infinite(),
-        value == 0.0,
-        &format!("{:.16e}", value.abs()),
-        17,
-        |precision| format!("{:.*e}", precision, value.abs()),
-        |candidate| candidate.parse::<f64>().ok().map(f64::to_bits) == Some(value.abs().to_bits()),
-    )
-}
-
-#[expect(
-    clippy::too_many_arguments,
-    clippy::fn_params_excessive_bools,
-    reason = "shared Java float formatting parameters"
-)]
-fn java_floating_string(
-    negative: bool,
-    nan: bool,
-    infinite: bool,
-    zero: bool,
-    precise: &str,
-    max_digits: usize,
-    scientific: impl Fn(usize) -> String,
-    rounds_to_value: impl Fn(&str) -> bool,
-) -> String {
-    if nan {
-        return "NaN".to_owned();
-    }
-    if infinite {
-        return if negative { "-Infinity" } else { "Infinity" }.to_owned();
-    }
-    if zero {
-        return if negative { "-0.0" } else { "0.0" }.to_owned();
-    }
-
-    let Some((precise_mantissa, _)) = precise.split_once('e') else {
-        panic!("Rust scientific formatting omitted its exponent");
-    };
-    let precise_digits = precise_mantissa.replace('.', "");
-    let one_digit_is_exact = precise_digits[1..].bytes().all(|digit| digit == b'0');
-    let mut selected = None;
-    for length in 1..=max_digits {
-        let formatted = scientific(length - 1);
-        let Some((mantissa, exponent)) = formatted.split_once('e') else {
-            panic!("Rust scientific formatting omitted its exponent");
-        };
-        if !rounds_to_value(&formatted) {
-            continue;
-        }
-        let Ok(exponent) = exponent.parse::<i32>() else {
-            panic!("Rust scientific formatting emitted a non-decimal exponent");
-        };
-        let digits = mantissa.replace('.', "");
-        selected = Some((digits, exponent - length as i32 + 1));
-        if length >= 2 || one_digit_is_exact {
-            break;
-        }
-    }
-    let Some((mut digits, decimal_exponent)) = selected else {
-        panic!("full-precision Rust decimal did not round-trip");
-    };
-    while digits.ends_with('0') {
-        digits.pop();
-    }
-    let scientific_exponent = digits.len() as i32 + decimal_exponent - 1;
-    let mut output = if (-3..0).contains(&scientific_exponent) {
-        format!(
-            "0.{}{}",
-            "0".repeat((-scientific_exponent - 1) as usize),
-            digits
-        )
-    } else if (0..7).contains(&scientific_exponent) {
-        let decimal_position = (scientific_exponent + 1) as usize;
-        if decimal_position >= digits.len() {
-            format!(
-                "{}{}.0",
-                digits,
-                "0".repeat(decimal_position - digits.len())
-            )
-        } else {
-            format!(
-                "{}.{}",
-                &digits[..decimal_position],
-                &digits[decimal_position..]
-            )
-        }
-    } else {
-        let fraction = if digits.len() == 1 { "0" } else { &digits[1..] };
-        format!("{}.{}E{scientific_exponent}", &digits[..1], fraction)
-    };
-    if negative {
-        output.insert(0, '-');
-    }
-    output
 }
